Fixes spelling of h_percent option.
Fixes issue with typo for h_percent option on installation. This caused
Neatline to use NULL for the default value for horizontal width of the Items
column on Neatline exhibits, so that column would not appear.

Adds tests to check for correct values for default Neatline options.

Removes misspelled option, and adds correction option, in upgrade check for
1.0.1. Updates existing upgrades for <= 1.0.6 to check for existence of
columns before altering tables, to avoid upgrade errors.